Open Distribution Fix Null Values
Modified the getters and setter functions for HR, SR, GR to include escape for NULL data type.

These fixes are dirty, it will pass 0 as an int in getter when value is null in table so these values will be displayed as that in editor.
For setter it will check if 65535 int was used (no limit) and change it to null before inserting into table.